archive/murcl-requirements.txt
torch
torch-geometric
torchvision
h5py
pandas
numpy
opencv-python
opencv-python-headless
scikit-learn
scikit-image
matplotlib
scipy
pyyaml
tensorboard
tqdm
pathtools
joblib
json-tricks
openslide-bin
openslide-python
sklearn-pandas
